Extract fixture.
